+++ libsolidity/interface/OptimiserSettings.h
@@ -34,9 +34,9 @@ struct OptimiserSettings
 	static char constexpr DefaultYulOptimiserSteps[] =
 		"dhfoDgvulfnTUtnIf"            // None of these can make stack problems worse
 		"["
-			"xarrscLM"                 // Turn into SSA and simplify
+			"xarrscLMS"                // Turn into SSA and simplify
 			"cCTUtTOntnfDIul"          // Perform structural simplification
-			"Lcul"                     // Simplify again
+			"LculS"                    // Simplify again
 			"Vcul jj"                  // Reverse SSA
 
 			// should have good "compilability" property here.